:bug: BUGFIX: Fix random labyrinth generation
Fix pseudo random generation. Added instructions to set the seed for the
random generation at each call to the system time so that actual
randomness can be achieved in the generated labyrinths.

Replace the random wall breaking algorithm with the one suggested in the
project paper to properly generate a random labyrinth with a unique
solution every time. This was retrofitted quick fix to get to working
state. Thus, I went slightly outside the limitations and rules in the
project paper.

Simplify `cell_mettre_bit_a0` function making use of the `not` operator.